pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["linroot <linroot@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
bs4 = "^0.0.1"
lxml = "^4.9.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
ruff = "^0.9.0"
jinja2 = "^3.1.2"
redis = "^5.0.1"
websockets = "^12.0"
wsproto = "^1.2.0"
alembic-utils = "^0.8.2"
python-multipart = "^0.0.9"
gunicorn = "^23.0.0"
pillow = "^10.3.0"
aiohttp = "^3.9.3"
pyyaml = "^6.0.2"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
keyring = "^25.0.1"
pylsp-rope = "^0.1.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.2"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-xdist = "^3.8.0"
pytest-cov = "^6.0.0"
pytest-benchmark = "^4.0.0"
hypothesis = "^6.115.0"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
